+++ osidb/tests/test_embargo.py
@@ -6,7 +6,7 @@ from osidb.models import Flaw
 
 from .factories import FlawFactory
 
-logger: logging = logging.getLogger(__name__)
+logger = logging.getLogger(__name__)
 pytestmark = pytest.mark.unit